Whitelist url with room persistence

+++ app/src/main/java/com/getadhell/androidapp/db/dao/AppInfoDao.java
@@ -28,7 +28,7 @@ public interface AppInfoDao
     void deleteAppInfoByPackageName(String packageName);
 
     @Query("SELECT MAX(id) FROM AppInfo")
-    int getMaxId();
+    long getMaxId();
 
     @Query("SELECT * FROM AppInfo ORDER BY appName ASC")
     List<AppInfo> getAll();
